# api/openapi.yaml
openapi: 3.0.3
info:
  title: Hasamex Transcript Analysis — Internal API
  description: >
    Internal REST API for the transcript analysis application.
    Consumed by the Streamlit frontend. All routes are versioned under /api/v1.
  version: 1.0.0
servers:
  - url: http://localhost:8000
paths:
...
  /api/v1/transcripts/upload:
    post:
      tags: [Transcripts]
      summary: Upload one or more transcript files
      description: >
        Multipart upload of .txt transcript files (max 5 MB each). Each file is parsed,
        embedded, and stored. Re-uploading an existing filename soft-deletes the older
        version and stores a new one (version increments); the interview-guide cache for
        that file is purged. Per-file results report status: uploaded | replaced | error.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Per-file upload results
          content:
            application/json:
              schema:
                type: object
                properties:
                  processed: { type: integer }
                  succeeded: { type: integer }
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        filename: { type: string }
                        status:
                          type: string
                          enum: [uploaded, replaced, error]
                        transcript_id: { type: integer }
                        version: { type: integer }
                        chunk_count: { type: integer }
                        reason: { type: string }

  # ══════════════════════════════════════════════════
  # Interview Guide Answers
  # ══════════════════════════════════════════════════
...
  /api/v1/qa/ask:
    post:
      tags: [Q&A]
      summary: Ask a free-form question across all transcripts
      description: >
        Single Q&A endpoint. Every question returns BOTH:

        1. a RAG answer — question embedded, top-K chunks retrieved via cosine
           similarity, grounded LLM answer with char-verified citations; and

        2. verbatim quotes — hybrid search (semantic + keyword) returns raw
           transcript chunks (no LLM), each marked verified with its timestamp,
           so claims can be checked word-for-word.

        When the question is off-topic or a greeting/capability message, the
        response is a conversational LLM scope reply (dedicated scope system
        prompt, free-text, cached per question) that explains what the interviews
        cover; canned messages are only the fallback if the LLM call fails.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question: { type: string, example: "What slows down adoption in Germany?" }
                top_k: { type: integer, default: 5 }
      responses:
        '200':
          description: LLM answer with citations PLUS verified verbatim quotes
          content:
            application/json:
              schema:
                type: object
                properties:
                  question: { type: string }
                  mode:
                    type: string
                    enum: [answer]
                  answer: { type: string }
                  quotes:
                    type: array
                    description: Verified verbatim chunks from the transcripts.
                    items:
                      type: object
                      properties:
                        quote: { type: string }
                        transcript_file: { type: string }
                        expert_name: { type: string }
                        market: { type: string }
                        timestamp: { type: string }
                        verification_status:
                          type: string
                          enum: [verified, verification_failed]
                  citations:
                    type: array
                    items:
                      type: object
                      properties:
                        transcript_file: { type: string }
                        expert_name: { type: string }
                        market: { type: string }
                        timestamp: { type: string }
                        quote: { type: string }